Update to python 3, migrate to flask (webapp2 incompatible with python 3)
