+++ python/vote-text-bot/README.md
@@ -2,9 +2,7 @@
 
 This code sample creates a Hangouts Chat simple text-only Vote bot that
 allows a user to vote (+1 or -1) on a message. This implementation is
-written in Python and hosted on Google App Engine. It uses the `webapp2`
-micro framework but can be ported to Flask if desired for App Engine or
-alternative hosting. It should look like this:
+written in Python and hosted on Google App Engine.
 
 ![Vote bot (text-only)](https://user-images.githubusercontent.com/744973/43745610-9fdfc1c4-9994-11e8-9f47-b16fe3438458.png)
