+++ python/vote-text-bot/app.yaml
@@ -1,7 +1,1 @@
-runtime: python27
-api_version: 1
-threadsafe: true
-
-handlers:
-- url: /.*
-  script: bot.app
+runtime: python37
