+++ python/vote-text-bot/requirements.txt
@@ -1,0 +1,1 @@
+Flask==1.1.1
